Refactor playwright to handle nondefault
Refactored playwright structure to allow for testing an instance with default config changed
- Added new compose file to test non-default settings
- Added new makefile targets and split by default and non-default
- Moved tests to default settings folder
